Fix miss match between completion.content.role_id type and completion.role.id

+++ databases/core_ai/tables/completion/content/create_table_completion_content.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS completion.content
     id bigserial PRIMARY KEY,
     message text,
     created_at timestamp DEFAULT timezone('UTC'::text, now()),
-    role_id bigint,
+    role_id smallint,
     account_id serial,
     persona_id serial
 );
